Fix UEFI boot issues
This adds a custom element to re-run grub-install without the
`--removeable` option, which causes it to correctly populate the
`/boot/efi/EFI/ubuntu` directory. This results in an image that will boot
successfully on UEFI systems.

You will run into an issue testing this image in a virtual machine: when
the image first boots, it immediately resets during the early UEFI boot
process. This will require you to `virsh start` the image, which will work
fine but will not attach cloud-init metadata you defined with the
`--cloud-init` option so you won't be able to log in.

Consider using `virt-customize` to set a root password before testing the
image.